Remove unnessessary parameter from InstallerAdapters uninstall()

+++ libraries/cms/installer/adapter/library.php
@@ -215,17 +215,15 @@ class JInstallerAdapterLibrary extends JInstallerAdapter
 	 * This method populates the $this->extension object, checks whether the extension is protected,
 	 * and sets the extension paths
 	 *
-	 * @param   integer  $id  The extension ID to load
-	 *
 	 * @return  boolean  True on success
 	 *
 	 * @note    Due to non-standard processing, the manifest is also set in this extended method
 	 * @since   3.1
 	 */
-	protected function setupUninstall($id)
+	protected function setupUninstall()
 	{
 		// Run the common parent methods
-		if (parent::setupUninstall($id))
+		if (parent::setupUninstall())
 		{
 			$this->manifestFile = JPATH_MANIFESTS . '/libraries/' . $this->element . '.xml';
 
@@ -275,18 +273,18 @@ class JInstallerAdapterLibrary extends JInstallerAdapter
 	/**
 	 * Custom uninstall method
 	 *
-	 * @param   string  $id  The id of the library to uninstall.
-	 *
 	 * @return  boolean  True on success
 	 *
 	 * @since   3.1
 	 */
-	public function uninstall($id)
+	public function uninstall()
 	{
-		$retval = true;
-
 		// Prepare the uninstaller for action
-		$this->setupUninstall((int) $id);
+		if (!$this->setupUninstall()) {
+			return false;
+		}
+
+		$retval = true;
 
 		$this->parent->removeFiles($this->manifest->files, -1);
 		JFile::delete($this->manifestFile);
